Fix performance bug
Face indices were combined in the wrong order - instead of adding (and offsetting) each new mesh's face indices to the accumulated mesh, for every new mesh the accumulated face indices were added to (and offsetted from) the new mesh. This led to O(n^2) instead of O(n) complexity in terms of number of meshes.